Flyttet ut typen for Stønadsperiode
